[hotfix] 예상 이익 int 형일때 억단위면 터지는 오류 bigdecimal 로 수정

+++ src/main/java/com/fiveguys/RIA/RIA_Backend/campaign/project/model/dto/request/ProjectUpdateRequestDto.java
@@ -18,7 +18,7 @@ public class ProjectUpdateRequestDto {
   private String title;                // 프로젝트명
   private String description;          // 설명
   private String type;                 // EXHIBITION / RENTAL / POPUP
-  private Integer expectedRevenue;     // 예상매출
+  private BigDecimal expectedRevenue;     // 예상매출
   private BigDecimal expectedMarginRate; // 예상이익률 (%)
   private LocalDate startDay;          // 시작일
   private LocalDate endDay;            // 종료일
